requirements.txt
requests
pandas
numpy
python-dotenv
slack-sdk